+++ AFTEREFFECT_PLUGIN.DEV/ImageLabProcLib/include/fft_czt.hpp
@@ -30,16 +30,16 @@ inline int32_t NextPowerOf2 (int32_t n) noexcept
 // This simplistic recursive implementation is here ONLY to make the CZT code 
 // below runnable in this snippet. Do not use this function in production.
 template <typename T>
-inline void FFT_Pow2_Helper (int32_t N, T* data, bool inverse) noexcept
+inline void FFT_Pow2_Helper (ptrdiff_t N, T* data, bool inverse) noexcept
 {
     if (N <= 1) return;
     
     // Separate even/odd
-    const int32_t M = N / 2;
+    const ptrdiff_t M = N / 2;
     std::vector<T> even(2 * M);
     std::vector<T> odd(2 * M);
     
-    for (int i = 0; i < M; ++i)
+    for (ptrdiff_t i = 0; i < M; ++i)
 	{
         even[2*i]   = data[4*i];   even[2*i+1] = data[4*i+1];
         odd[2*i]    = data[4*i+2]; odd[2*i+1] = data[4*i+3];
@@ -51,7 +51,7 @@ inline void FFT_Pow2_Helper (int32_t N, T* data, bool inverse) noexcept
     constexpr T PI {static_cast<T>(3.14159265358979323846)};
     const T angle_dir = inverse ? static_cast<T>(2.0) * PI : static_cast<T>(-2.0) * PI;
     
-    for (int32_t k = 0; k < M; ++k)
+    for (ptrdiff_t k = 0; k < M; ++k)
 	{
         // W = exp(j * angle * k / N)
         const T theta = angle_dir * static_cast<T>(k) / static_cast<T>(N);
@@ -78,7 +78,7 @@ inline void FFT_Pow2_Helper (int32_t N, T* data, bool inverse) noexcept
 // POWER-OF-2 FFT INTERFACE (Use your engine here)
 // ----------------------------------------------------------------------------
 template <typename T>
-inline void Execute_FFT_Pow2 (int32_t M, T* data) noexcept
+inline void Execute_FFT_Pow2 (ptrdiff_t M, T* data) noexcept
 {
     // In your real code: Execute_FFT_Iterative(M, data, {16, 16, 4...});
     FFT_Pow2_Helper (M, data, false);
@@ -92,7 +92,7 @@ inline void Execute_IFFT_Pow2 (int32_t M, T* data) noexcept
     
     // Normalize
     const T invM = static_cast<T>(1) / static_cast<T>(M);
-    for (int i = 0; i < 2 * M; ++i) data[i] *= invM;
+    for (ptrdiff_t i = 0; i < 2 * M; ++i) data[i] *= invM;
 }
 
 // ============================================================================
@@ -124,7 +124,7 @@ inline void fft_czt (const T* in, T* out, int32_t N) noexcept
     // The Chirp is W_k = exp( -j * PI * k^2 / N )
     // Note: The formula uses PI, not 2*PI, due to the n^2/2 identity.
     
-    for (int k = 0; k < N; ++k)
+    for (ptrdiff_t k = 0; k < N; ++k)
 	{
         // Calculate angle: -PI * k^2 / N
         T k_sq = static_cast<T>(1 * k * k); // 1LL prevents overflow
@@ -158,7 +158,7 @@ inline void fft_czt (const T* in, T* out, int32_t N) noexcept
 
     // Negative Lags (Wrap to end of M)
     // B[M-k] = B[k] for k = 1..N-1
-    for (int32_t k = 1; k < N; ++k)
+    for (ptrdiff_t k = 1; k < N; ++k)
 	{
         int32_t wrap_idx = M - k;
         B[2*wrap_idx]     = chirp[2*k];     // Real
@@ -175,7 +175,7 @@ inline void fft_czt (const T* in, T* out, int32_t N) noexcept
 
     // B. Pointwise Multiplication in Frequency Domain
     // A = A * B
-    for (int32_t k = 0; k < M; ++k)
+    for (ptrdiff_t k = 0; k < M; ++k)
 	{
         T ar = A[2*k];     T ai = A[2*k + 1];
         T br = B[2*k];     T bi = B[2*k + 1];
@@ -195,7 +195,7 @@ inline void fft_czt (const T* in, T* out, int32_t N) noexcept
     // Output[k] = A[k] * Chirp[k]
     // Only the first N elements are valid.
 
-    for (int32_t k = 0; k < N; ++k)
+    for (ptrdiff_t k = 0; k < N; ++k)
 	{
         T ar = A[2*k];
         T ai = A[2*k + 1];
